# server/src/RestApi/openapi.yaml
openapi: '3.1.1'
info:
  title: Metaspesa REST API
  description: Errors use ProblemDetails with a stable code and traceId for support.
  version: 1.0.0
servers:
  - url: /
paths:
  /api/v1/auth/registrations:
    post:
      tags:
        - RegisterEndpoint
      summary: Register a shopper.
      description: Create a shopper account. Requests must come from an allowed browser origin.
      operationId: RegisterShopper
      requestBody:
        description: Credentials supplied when registering or authenticating a shopper.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsRequest'
        required: true
      responses:
        '201':
          description: Account created; no response body.
        '400':
          description: 'Username or password violates registration rules, or the body is invalid.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: The username is already registered.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The browser Origin header is missing or not allowed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: An unexpected server or database failure occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/sessions:
    post:
      tags:
        - CreateSessionEndpoint
      summary: Create a browser session.
      description: Authenticate a shopper and set an HTTP-only session cookie. Requests must come from an allowed browser origin.
      operationId: CreateBrowserSession
      requestBody:
        description: Credentials supplied when registering or authenticating a shopper.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsRequest'
        required: true
      responses:
        '204':
          description: Authenticated. A session cookie is set; no response body.
        '400':
          description: The request body is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: The username or password is incorrect.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The browser Origin header is missing or not allowed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: An unexpected server or database failure occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/tokens:
    post:
      tags:
        - CreateTokenEndpoint
      summary: Create a machine token.
      description: Authenticate credentials and return a bearer token for non-browser clients.
      operationId: CreateMachineToken
      requestBody:
        description: Credentials supplied when registering or authenticating a shopper.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsRequest'
        required: true
      responses:
        '200':
          description: Authenticated. Returns a bearer token and its expiry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: The request body is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: The username or password is incorrect.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: An unexpected server or database failure occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/markets:
    get:
      tags:
        - GetMarketsEndpoint
      summary: List registered markets.
      description: Return markets available for catalog filtering and product display.
      operationId: GetMarkets
      responses:
        '200':
          description: Available markets and their stable IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketListResponse'
        '500':
          description: An unexpected server or database failure occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/products:
    get:
      tags:
        - GetProductsEndpoint
      summary: Search catalog products.
      description: 'Search market-owned products with optional filters. Each result includes formats with their latest observed prices. Pagination counts products, not formats.'
      operationId: GetProducts
      parameters:
        - name: query
          in: query
          description: Case-insensitive product-name fragment. Omit to include all names.
          schema:
            type: string
        - name: marketId
          in: query
          description: Repeat to include products from several markets. IDs must be positive.
          schema:
            type: array
            items:
              minimum: 1
              type: integer
        - name: brand
          in: query
          description: Case-insensitive brand-name fragment. Omit to include all brands.
          schema:
            type: string
        - name: page
          in: query
          description: One-based page number; defaults to 1.
          schema:
            minimum: 1
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: 'Maximum products per page, from 1 to 100; defaults to 24.'
          schema:
            maximum: 100
            minimum: 1
            type: integer
            default: 24
        - name: sort
          in: query
          description: Sort by name or lowest latest format price. Defaults to name.
          schema:
            enum:
              - name
              - priceAsc
              - priceDesc
            type: string
      responses:
        '200':
          description: Matching products and pagination counts. An empty page has no items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '400':
          description: 'A filter, market ID, page value, or sort value is invalid.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: An unexpected server or database failure occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/v1/markets/{marketName}/snapshots/{date}':
    post:
      tags:
        - SnapshotEndpoint
      summary: Import a market price snapshot.
      description: Import product observations for a market and date. Requires authentication with the ProductManager role.
      operationId: AddMarketSnapshot
      parameters:
        - name: marketName
          in: path
          description: Name of the observed market.
          required: true
          schema:
            type: string
        - name: date
          in: path
          description: 'Observation date in YYYY-MM-DD format, on or after 2023-01-01.'
          required: true
          schema:
            type: string
      requestBody:
        description: Batch of market product observations.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotRequest'
        required: true
      responses:
        '204':
          description: Snapshot imported; no response body.
        '400':
          description: The date or a product observation is invalid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: No valid authentication token was provided.
        '403':
          description: The authenticated user lacks the ProductManager role.
        '500':
          description: An unexpected server or database failure occurred.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    CredentialsRequest:
      required:
        - username
        - password
      type: object
      properties:
        username:
          type: string
          description: Shopper username used to identify the account.
        password:
          type: string
          description: Plaintext password for this request; never returned in responses.
      description: Credentials supplied when registering or authenticating a shopper.
    FormatResponse:
      required:
        - id
        - quantity
        - currentPrice
        - observedAt
      type: object
      properties:
        id:
          type: integer
          description: Stable format ID used as the shopping-list item reference.
          format: int32
        quantity:
          description: Amount and unit sold in this format.
          $ref: '#/components/schemas/QuantityResponse'
        currentPrice:
          description: Latest observed price for this format.
          $ref: '#/components/schemas/MoneyResponse'
        observedAt:
          type: string
          description: UTC instant of the latest price observation.
          format: date-time
        imageUrl:
          type:
            - 'null'
            - string
          description: Absolute product image URL when available; otherwise omitted.
      description: Purchasable product format with its latest price observation.
    MarketListResponse:
      required:
        - items
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MarketResponse'
          description: Market summaries; empty when no markets are registered.
      description: Markets available in the catalog.
    MarketResponse:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          description: Stable market ID used by the catalog marketId filter.
          format: int32
        name:
          type: string
          description: Display name of the market.
        logoUrl:
          type:
            - 'null'
            - string
          description: Absolute logo URL when available; otherwise omitted.
      description: Market identity and display information.
    MoneyResponse:
      required:
        - amount
        - currency
      type: object
      properties:
        amount:
          type: number
          description: Price in the stated currency.
          format: double
        currency:
          type: string
          description: 'Currency code of the price, such as EUR.'
      description: Monetary amount and currency.
    ProblemDetails:
      required:
        - code
        - traceId
      type: object
      properties:
        type:
          type:
            - 'null'
            - string
          description: URI identifying the class of problem.
        title:
          type:
            - 'null'
            - string
          description: Short human-readable reason for the failure.
        status:
          type:
            - 'null'
            - integer
          description: HTTP status code of this response.
          format: int32
        detail:
          type:
            - 'null'
            - string
          description: Additional detail when available.
        instance:
          type:
            - 'null'
            - string
          description: Request URI when available.
        code:
          type: string
          description: Stable machine-readable error code.
        traceId:
          type: string
          description: Request trace identifier for support and logs.
      description: Error response with a machine-readable code and request trace identifier.
    ProductPageResponse:
      required:
        - items
        - page
        - pageSize
        - totalItems
        - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
          description: Products in this page; empty when none match.
        page:
          type: integer
          description: One-based page number requested.
          format: int32
        pageSize:
          type: integer
          description: Maximum number of products requested for this page.
          format: int32
        totalItems:
          type: integer
          description: Number of matching products across all pages.
          format: int32
        totalPages:
          type: integer
          description: Number of pages at this page size; zero when none match.
          format: int32
      description: One page of matching products.
    ProductResponse:
      required:
        - id
        - name
        - brand
        - market
        - formats
      type: object
      properties:
        id:
          type: integer
          description: Stable ID of this product.
          format: int32
        name:
          type: string
          description: Product display name.
        brand:
          type: string
          description: Product brand display name.
        market:
          description: Market that owns this product.
          $ref: '#/components/schemas/MarketResponse'
        formats:
          type: array
          items:
            $ref: '#/components/schemas/FormatResponse'
          description: Available quantities and their latest observed prices.
      description: One market-owned product. IDs do not imply cross-market equivalence.
    QuantityResponse:
      required:
        - amount
        - unit
      type: object
      properties:
        amount:
          type: number
          description: Numeric quantity of the format.
          format: double
        unit:
          type: string
          description: 'Unit of measure, such as kg, g, or l.'
      description: Quantity associated with one product format.
    SnapshotItem:
      required:
        - name
        - price
        - quantity
        - unitOfMeasure
        - brandName
        - imageUrl
      type: object
      properties:
        name:
          type:
            - 'null'
            - string
          description: Required nonblank product name.
        price:
          type: number
          description: Observed non-negative price amount.
          format: double
        quantity:
          type: number
          description: Positive amount sold in this format.
          format: float
        unitOfMeasure:
          type:
            - 'null'
            - string
          description: Unit of the quantity; must be supported.
        brandName:
          type:
            - 'null'
            - string
          description: Required nonblank brand name.
        imageUrl:
          type:
            - 'null'
            - string
          description: Optional absolute product image URL.
      description: One observed product format in a market snapshot.
    SnapshotRequest:
      required:
        - items
      type: object
      properties:
        items:
          type:
            - 'null'
            - array
          items:
            $ref: '#/components/schemas/SnapshotItem'
          description: Non-empty array of valid product observations.
      description: Batch of market product observations.
    TokenResponse:
      required:
        - accessToken
        - tokenType
        - expiresAt
      type: object
      properties:
        accessToken:
          type: string
          description: JWT sent in the Authorization Bearer header.
        tokenType:
          type: string
          description: Authorization scheme; always Bearer.
        expiresAt:
          type: string
          description: UTC instant after which the token is invalid.
          format: date-time
      description: Bearer token issued after successful machine authentication.
tags:
  - name: RegisterEndpoint
  - name: CreateSessionEndpoint
  - name: CreateTokenEndpoint
  - name: GetMarketsEndpoint
  - name: GetProductsEndpoint
  - name: SnapshotEndpoint